requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
decorator==5.1.1
Django==4.0.4
django-crispy-forms==1.14.0
django-localflavor==3.1
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geopy==2.2.0
idna==3.3
python-stdnum==1.17
PyYAML==6.0
ratelim==0.1.6
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
